requirements.txt
amqp==5.1.1
asgiref==3.2.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.7.0
backports.zoneinfo==0.2.1
bandit==1.7.4
billiard==3.6.4.0
bleach==4.1.0
boto3==1.23.5
botocore==1.26.5
CacheControl==0.12.6
cachetools==4.2.4
celery==5.1.2
certifi==2022.9.24
cffi==1.14.4
chardet==4.0.0
charset-normalizer==2.1.1
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.6.0
django-extensions==3.1.5
django-haystack==3.1.1
django-logic==0.1.2
django-oauth-toolkit==2.1.0
django-shell-plus==1.1.7
django-showurls==1.0.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.8.0
drf-haystack==1.8.11
drf-social-oauth2==1.0.9
et-xmlfile==1.1.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.28
gunicorn==20.0.4
idna==2.10
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==1.0.1
jwcrypto==1.4.2
kombu==5.2.4
MarkupSafe==2.1.1
msgpack==1.0.4
oauthlib==3.2.1
openpyxl==3.0.9
packaging==21.3
pbr==5.10.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
smmap==5.0.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stevedore==4.0.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.12
uuid==1.30
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
